requirements.txt
click==8.1.7
openai==1.12.0
python-dotenv==1.0.1
rich==13.7.0
tree-sitter==0.20.4
pathlib==1.0.1 